+++ src/headers/Sphere.h
@@ -5,8 +5,8 @@
 #include "Geometry.h"
 class Sphere:public Geometry{
 	public:
-	float radius;
-	Sphere(float radius=1)noexcept:radius(radius){}
+	double radius;
+	Sphere(double radius=1)noexcept:radius(radius){}
 	IntersectInfo testIntersect(Ray r)const noexcept override;
 };
 #endif //SPHERE_H
